requirements.txt
# Core dependencies
langgraph>=0.0.10
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyyaml>=6.0
docker>=6.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Optional
pytest>=7.0 ; python_version >= "3.9"
black>=22.0 ; python_version >= "3.9"
flake8>=5.0 ; python_version >= "3.9"
